Make the SQLAlchemy model / database schema comparison reusable
Move the comparison itself into inmanta.db.schema_compare, so an extension
can verify its own models against the schema its own migration scripts
produce. The test in tests/db keeps the drift cases, which now exercise the
shared code, and the comparison against the real database.
